rework: centralize helper function
