+++ backend/lib/logger/node/index.ts
@@ -1,6 +1,6 @@
 import { IRuntime } from "../../types/index.js";
 import { getMeta, transportJSON, transportFormatted } from "./runtime.js";
-import { getCallerStackFrame, getErrorTrace, prettyFormatLogObj, prettyFormatErrorObj, isError, isBuffer } from "./util.js";
+import { getCallerStackFrame, getErrorTrace, prettyFormatLogObj, prettyFormatErrorObj, isError, isBuffer } from "../../helper/index.js";
 
 export default {
     getCallerStackFrame,
