+++ backend/lib/types/index.d.ts
@@ -4,3 +4,4 @@ export * from './forwarder.d.ts';
 export * from './logger.d.ts';
 export * from './node.d.ts';
 export * from './middleware.d.ts';
+export * from './object-merge.d.ts';
